Add Junit Test without gradle
