+++ src/main/java/test/Deadlines.java
@@ -1,15 +1,17 @@
+package test;
 import java.time.LocalDate;
+
 import java.time.format.DateTimeFormatter;
 import java.time.temporal.ChronoUnit;
 
-public class Events extends Task{
+public class Deadlines extends test.Task{
     private LocalDate date;
 
-    public Events(String newDescripton, String newTime){
+    public Deadlines(String newDescripton, String newTime){
         super(newDescripton);
         this.date = LocalDate.parse(newTime);
     }
-    public Events(String newDescripton, String newTime, String status){
+    public Deadlines(String newDescripton, String newTime, String status){
         super(newDescripton, status);
         this.date = LocalDate.parse(newTime);
     }
@@ -18,15 +20,16 @@ public class Events extends Task{
                 + this.date.getMonth().toString() + " "
                 + Integer.valueOf(this.date.getYear()).toString() ;
     }
+
     public LocalDate getDate() {
         return date;
     }
 
     @Override
     public String toString(){
-        return " [E][" + super.getStatusIcon() + "]" + super.description + " "
-               + Integer.valueOf(this.date.getDayOfMonth()).toString() + " "
-                 + this.date.getMonth().toString() + " "
-                    + Integer.valueOf(this.date.getYear()).toString() ;
+        return " [D][" + super.getStatusIcon() + "]" + super.description + " "
+                + Integer.valueOf(this.date.getDayOfMonth()).toString() + "-"
+                    + this.date.getMonth().toString() + "-"
+                        + Integer.valueOf(this.date.getYear()).toString() ;
     }
 }
