+++ src/main/java/test/Task.java
@@ -1,3 +1,4 @@
+package test;
 public class Task {
     protected String description;
     protected boolean isDone;
@@ -19,6 +20,7 @@ public class Task {
         return this;
     }
     public String checkDone(){
+
         return(isDone ? "Yes" : "No");
     }
